+++ lib/my/my_math/my_putnbr_base.c
@@ -8,7 +8,7 @@
 #include <my_stdio.h>
 #include <stdbool.h>
 
-static void process_number(int nb, int base_size, char const *base)
+static void process_number(long long nb, int base_size, char const *base)
 {
     if (nb >= base_size) {
         process_number(nb / base_size, base_size, base);
@@ -17,7 +17,7 @@ static void process_number(int nb, int base_size, char const *base)
         my_putchar(base[nb]);
 }
 
-void my_putnbr_base(int nb, int base_size, bool uppercase)
+void my_putnbr_base(long long nb, int base_size, bool uppercase)
 {
     char const base_higher[36] = "0123456789ABCDEFGHJKLMNOPQRSTUVWXZ";
     char const base_lower[36] = "0123456789abcdefghijklmnopqrstuvwxyz";
